+++ src/DocSharp.Markdown/Docx/Blocks/ListRenderer.cs
@@ -25,12 +25,12 @@ public class ListRenderer : DocxObjectRenderer<ListBlock>
     {
         var listInfo = new ListInfo();
         var numbering = renderer.Document.GetOrCreateNumbering().NumberingDefinitionsPart!.Numbering;
-        var listStyle = obj.IsOrdered ? renderer.Styles.MarkdownStyles["ListOrdered"] : renderer.Styles.MarkdownStyles["ListBullet"];
-        var listItemStyle = obj.IsOrdered ? renderer.Styles.MarkdownStyles["ListOrderedItem"] : renderer.Styles.MarkdownStyles["ListBulletItem"];
+        var listItemStyle = obj.IsOrdered ? renderer.Styles.MarkdownStyles["ListOrderedItem"] : 
+                                            renderer.Styles.MarkdownStyles["ListBulletItem"];
         listInfo.StyleId = listItemStyle;
 
-        var abstractNum = numbering.Elements<AbstractNum>().FirstOrDefault(e => e.StyleLink?.Val == listStyle);
-        if (abstractNum?.AbstractNumberId != null) // TODO: Fallback and create this
+        var abstractNum = numbering.Elements<AbstractNum>().FirstOrDefault(e => e.StyleLink?.Val == listItemStyle);
+        if (abstractNum?.AbstractNumberId != null) // TODO: Fallback and create this; or try to get try to get style --> numPr --> numId --> abstractNumId instead.
         {
             int abstractNumId = abstractNum.AbstractNumberId.Value;
 
@@ -38,7 +38,7 @@ public class ListRenderer : DocxObjectRenderer<ListBlock>
             //                                                                     n.AbstractNumId.Val == abstractNumId)
             //                                                         .FirstOrDefault();
 
-            var newNumberingId = numbering.Elements<NumberingInstance>().Count() + 1;
+            var newNumberingId = numbering.Elements<NumberingInstance>().Select(n => n.NumberID?.Value ?? 0).DefaultIfEmpty(0).Max() + 1;
             var numberingInstance = new NumberingInstance
             {
                 NumberID = newNumberingId,